requirements.txt
PyQt5
geopandas
shapely
fiona
pyproj
google-generativeai
pyyaml
python-dotenv
matplotlib
folium
PyQtWebEngine
QDarkStyle
